requirements.txt
numpy>=1.13.0
scipy>=0.19.1
pytest-cov>=2.5.1
